fix: cast .mat inputs to double, matching the NIfTI branch
qMRLab's voxelwise fits fail at EVERY voxel on non-double input, and the
failure is close to invisible: the error reporter (cprintf) falls over
while reporting it, so what surfaces is a crash in a printing utility
rather than a type problem in the data. Diagnosed last round: qmt_spgr's
MTdata.mat/Mask.mat archive ships single precision, lsqcurvefit inside
qmt_spgr's fit rejects it outright (optimlib:lsqcurvefit:NonDoubleInput),
and qMRLab's own error-reporting path (FitData -> cprintf) then crashes
on an unescaped quote in the message it's trying to print, masking the
real cause behind a MATLAB:m_missing_operator parse error.

Verified with real MATLAB R2026a + qMRLab v3.0.0, cprintf completely
unmodified (`which cprintf` resolves to qMRLab's own bundled copy):
qmt_spgr now reports ok and matches its archive close to float precision
(F: frac_within=0.9976, pearson_r=0.9999998541730009; kr/R1f/T2f/T2r
similarly >0.995/>0.9999; R1r is a fixed parameter, exact match, pearson
undefined due to zero variance). No cprintf shim anywhere in the repo.

python -m pytest tests/ -v: 81 passed.

+++ matlab/qmrlab_ci_load_any.m
@@ -10,7 +10,11 @@ function arr = qmrlab_ci_load_any(dataRoot, dataset, name)
     if endsWith(name, '.mat')
         s = load(p);
         f = fieldnames(s);
-        arr = s.(f{1});
+        % double(), matching the NIfTI branch. qMRLab's voxelwise fits fail at EVERY
+        % voxel on non-double input, and the failure is close to invisible: the error
+        % reporter (cprintf) falls over while reporting it, so what surfaces is a crash
+        % in a printing utility rather than a type problem in the data.
+        arr = double(s.(f{1}));
     else
         arr = double(load_nii_data(p));
     end
